requirements.txt
streamlit
googletrans==4.0.0-rc1
PyPDF2
python-docx
pandas
openpyxl
reportlab